add real file handles to wrap BPTRs

+++ newlib/libc/sys/amigaos/include/sys/reent.h
@@ -186,7 +186,7 @@ struct __sFILE {
   int	_r;		/* read space left for getc() */
   int	_w;		/* write space left for putc() */
   short	_flags;		/* flags, below; this FILE is free if 0 */
-  int	_file;		/* fileno, if Unix descriptor, else -1 */
+  short	_file;		/* fileno, if Unix descriptor, else -1 */
   struct __sbuf _bf;	/* the buffer (at least 1 byte, if !NULL) */
   int	_lbfsize;	/* 0 or -_bf._size, for inline putc */
